change data type as float

+++ database/migrations/2021_12_06_160048_create_lands_table.php
@@ -18,7 +18,7 @@ class CreateLandsTable extends Migration
             $table->string('landAddress');
             $table->string('landMap');
             $table->string('landDetails');
-            $table->BigInteger('landValue');
+            $table->float('landValue',10,2);
 
             $table->integer('ownerID');
